+++ app/src/main/java/org/breezyweather/sources/ipsb/IpSbLocationService.kt
@@ -29,7 +29,6 @@ import java.util.TimeZone
 import javax.inject.Inject
 
 class IpSbLocationService @Inject constructor(
-    @ApplicationContext context: Context,
     client: Retrofit.Builder
 ) : HttpSource(), LocationSource {